[TASK] Repair Unit- and FunctionalTests
Resolves: #216
