+++ Tests/Functional/Service/EventServiceTest.php
@@ -10,6 +10,7 @@
 namespace JWeiland\Events2\Tests\Unit\Service;
 
 use JWeiland\Events2\Configuration\ExtConf;
+use JWeiland\Events2\Domain\Factory\TimeFactory;
 use JWeiland\Events2\Domain\Model\Day;
 use JWeiland\Events2\Domain\Model\Event;
 use JWeiland\Events2\Domain\Model\Exception;
@@ -50,8 +51,7 @@ class EventServiceTest extends FunctionalTestCase
         $this->subject = GeneralUtility::makeInstance(
             EventService::class,
             $this->eventRepositoryProphecy->reveal(),
-            GeneralUtility::makeInstance(ExtConf::class),
-            GeneralUtility::makeInstance(DateTimeUtility::class)
+            new TimeFactory()
         );
     }
 
@@ -63,607 +63,6 @@ class EventServiceTest extends FunctionalTestCase
         );
 
         parent::tearDown();
-    }
-
-    /**
-     * @test
-     */
-    public function getExceptionsForDateReturnZeroExceptions()
-    {
-        $event = new Event();
-
-        self::assertEquals(
-            new \SplObjectStorage(),
-            $this->subject->getExceptionsForDate($event, new \DateTime())
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getExceptionsForDateWithRemoveExceptionReturnsZeroExceptionsForAdd()
-    {
-        $date = new \DateTime('midnight');
-
-        $exception = new Exception();
-        $exception->setExceptionType('Remove');
-        $exception->setExceptionDate($date);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($exception);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        self::assertEquals(
-            new \SplObjectStorage(),
-            $this->subject->getExceptionsForDate($event, $date, 'Add')
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getExceptionsForDateWithRemoveExceptionReturnsOneRemoveException()
-    {
-        $date = new \DateTime('midnight');
-
-        $exception = new Exception();
-        $exception->setExceptionType('Remove');
-        $exception->setExceptionDate($date);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($exception);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedExceptions = new \SplObjectStorage();
-        $expectedExceptions->attach($exception);
-
-        self::assertEquals(
-            $expectedExceptions,
-            $this->subject->getExceptionsForDate($event, $date, 'Remove')
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getExceptionsForDateWithRemoveExceptionWithNonNormalizedDateReturnsOneRemoveException()
-    {
-        $date = new \DateTime('now'); // date must be sanitized to midnight in getExceptionsForDate
-
-        $exception = new Exception();
-        $exception->setExceptionType('Remove');
-        $exception->setExceptionDate($date);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($exception);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedExceptions = new \SplObjectStorage();
-        $expectedExceptions->attach($exception);
-
-        self::assertEquals(
-            $expectedExceptions,
-            $this->subject->getExceptionsForDate($event, $date, 'Remove')
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getExceptionsForDateWithDifferentExceptionsReturnsAddException()
-    {
-        $date = new \DateTime('midnight');
-
-        $removeException = new Exception();
-        $removeException->setExceptionType('Remove');
-        $removeException->setExceptionDate($date);
-        $addException = new Exception();
-        $addException->setExceptionType('Add');
-        $addException->setExceptionDate($date);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($removeException);
-        $exceptions->attach($addException);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedAddExceptions = new \SplObjectStorage();
-        $expectedAddExceptions->attach($addException);
-
-        self::assertEquals(
-            $expectedAddExceptions,
-            $this->subject->getExceptionsForDate($event, $date, 'Add')
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getExceptionsForDateWithExceptionsOfDifferentDatesReturnsAddException()
-    {
-        $firstDate = new \DateTime('midnight');
-        $secondDate = new \DateTime('midnight');
-        $secondDate->modify('tomorrow');
-
-        $firstAddException = new Exception();
-        $firstAddException->setExceptionType('Add');
-        $firstAddException->setExceptionDate($firstDate);
-        $secondAddException = new Exception();
-        $secondAddException->setExceptionType('Add');
-        $secondAddException->setExceptionDate($secondDate);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($firstAddException);
-        $exceptions->attach($secondAddException);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedAddExceptions = new \SplObjectStorage();
-        $expectedAddExceptions->attach($firstAddException);
-
-        self::assertEquals(
-            $expectedAddExceptions,
-            $this->subject->getExceptionsForDate($event, $firstDate, 'Add')
-        );
-    }
-
-    /**
-     * This test also checks against lowercased and multiple spaces in list of exception types
-     *
-     * @test
-     */
-    public function getExceptionsForDateWithExceptionsOfDifferentDatesReturnsDifferentExceptions()
-    {
-        $firstDate = new \DateTime('midnight');
-        $secondDate = new \DateTime('midnight');
-        $secondDate->modify('tomorrow');
-
-        $firstAddException = new Exception();
-        $firstAddException->setExceptionType('Add');
-        $firstAddException->setExceptionDate($firstDate);
-        $secondAddException = new Exception();
-        $secondAddException->setExceptionType('Add');
-        $secondAddException->setExceptionDate($secondDate);
-        $timeException = new Exception();
-        $timeException->setExceptionType('Time');
-        $timeException->setExceptionDate($firstDate);
-        $infoException = new Exception();
-        $infoException->setExceptionType('Info');
-        $infoException->setExceptionDate($firstDate);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($firstAddException);
-        $exceptions->attach($secondAddException);
-        $exceptions->attach($timeException);
-        $exceptions->attach($infoException);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedExceptions = new \SplObjectStorage();
-        $expectedExceptions->attach($firstAddException);
-        $expectedExceptions->attach($timeException);
-        $expectedExceptions->attach($infoException);
-
-        self::assertEquals(
-            $expectedExceptions,
-            $this->subject->getExceptionsForDate($event, $firstDate, 'add, time,  info')
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithoutAnyTimesReturnsNoTimes()
-    {
-        $firstDate = new \DateTime('midnight');
-        $secondDate = new \DateTime('midnight');
-        $secondDate->modify('tomorrow');
-
-        $firstAddException = new Exception();
-        $firstAddException->setExceptionType('Add');
-        $firstAddException->setExceptionDate($firstDate);
-        $secondAddException = new Exception();
-        $secondAddException->setExceptionType('Add');
-        $secondAddException->setExceptionDate($secondDate);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($firstAddException);
-        $exceptions->attach($secondAddException);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedTimes = new \SplObjectStorage();
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithExceptionsWithTimeReturnsTimes()
-    {
-        $firstDate = new \DateTime('midnight');
-        $secondDate = new \DateTime('midnight');
-        $secondDate->modify('tomorrow');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-
-        $firstAddException = new Exception();
-        $firstAddException->setExceptionType('Add');
-        $firstAddException->setExceptionDate($firstDate);
-        $firstAddException->setExceptionTime($time);
-        $secondAddException = new Exception();
-        $secondAddException->setExceptionType('Add');
-        $secondAddException->setExceptionDate($secondDate);
-        $secondAddException->setExceptionTime($time);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($firstAddException);
-        $exceptions->attach($secondAddException);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithDifferentExceptionTypesWithTimeReturnsTimes()
-    {
-        $firstDate = new \DateTime('midnight');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-
-        $firstAddException = new Exception();
-        $firstAddException->setExceptionType('Add');
-        $firstAddException->setExceptionDate($firstDate);
-        $firstAddException->setExceptionTime($time);
-        $secondAddException = new Exception();
-        $secondAddException->setExceptionType('Info');
-        $secondAddException->setExceptionDate($firstDate);
-        $secondAddException->setExceptionTime($time);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($firstAddException);
-        $exceptions->attach($secondAddException);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time);
-        $expectedTimes->attach($time);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithDifferentTimesAndSingleEventReturnsNoTimes()
-    {
-        $firstDate = new \DateTime('midnight');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-
-        $times = new ObjectStorage();
-        $times->attach($time);
-
-        $event = new Event();
-        $event->setEventType('single');
-        $event->setDifferentTimes($times);
-
-        $expectedTimes = new \SplObjectStorage();
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithDifferentTimesOnDifferentWeekdayAndRecurringEventReturnsNoTimes()
-    {
-        $firstDate = new \DateTime('midnight');
-        $secondDate = new \DateTime('midnight');
-        $secondDate->modify('tomorrow');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-        $time->setWeekday($secondDate->format('l'));
-
-        $times = new ObjectStorage();
-        $times->attach($time);
-
-        $event = new Event();
-        $event->setEventType('recurring');
-        $event->setDifferentTimes($times);
-
-        $expectedTimes = new \SplObjectStorage();
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithDifferentTimesAndRecurringEventReturnsTimes()
-    {
-        $firstDate = new \DateTime('midnight');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-        $time->setWeekday($firstDate->format('l'));
-
-        $times = new ObjectStorage();
-        $times->attach($time);
-
-        $event = new Event();
-        $event->setEventType('recurring');
-        $event->setDifferentTimes($times);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithDifferentTimesAndSingleEventReturnsOneTime()
-    {
-        $firstDate = new \DateTime('midnight');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-
-        $times = new ObjectStorage();
-        $times->attach($time);
-
-        $event = new Event();
-        $event->setEventType('single');
-        $event->setDifferentTimes($times);
-        $event->setEventTime($time);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithEventTimeReturnsOneTime()
-    {
-        $firstDate = new \DateTime('midnight');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-
-        $event = new Event();
-        $event->setEventType('single');
-        $event->setEventTime($time);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTimesForDateWithExceptionsWithoutTimesButWithEventTimeReturnsOneTime()
-    {
-        $firstDate = new \DateTime('midnight');
-        $secondDate = new \DateTime('midnight');
-        $secondDate->modify('tomorrow');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-
-        $firstAddException = new Exception();
-        $firstAddException->setExceptionType('Add');
-        $firstAddException->setExceptionDate($firstDate);
-        $secondAddException = new Exception();
-        $secondAddException->setExceptionType('Add');
-        $secondAddException->setExceptionDate($secondDate);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($firstAddException);
-        $exceptions->attach($secondAddException);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-        $event->setEventTime($time);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getTimesForDate($event, $firstDate)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getSortedTimesForDateWithoutTimesReturnsEmptyStorage()
-    {
-        $date = new \DateTime('midnight');
-
-        $event = new Event();
-
-        $expectedTimes = new \SplObjectStorage();
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getSortedTimesForDate($event, $date)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getSortedTimesForDateWithEventTimeReturnsOneSortedTime()
-    {
-        $date = new \DateTime('midnight');
-
-        $time = new Time();
-        $time->setTimeBegin('10:30');
-
-        $event = new Event();
-        $event->setEventTime($time);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getSortedTimesForDate($event, $date)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getSortedTimesForDateWithDifferentTimesReturnsSortedTimes()
-    {
-        $date = new \DateTime('midnight');
-        $tomorrow = new \DateTime('midnight');
-        $tomorrow->modify('tomorrow');
-
-        $time1 = new Time();
-        $time1->setWeekday($date->format('l'));
-        $time1->setTimeBegin('23:56');
-        $time2 = new Time();
-        $time2->setWeekday($date->format('l'));
-        $time2->setTimeBegin('12:34');
-        $time3 = new Time();
-        $time3->setWeekday($date->format('l'));
-        $time3->setTimeBegin('00:34');
-        $time4 = new Time();
-        $time4->setWeekday($tomorrow->format('l'));
-        $time4->setTimeBegin('10:35');
-
-        $times = new ObjectStorage();
-        $times->attach($time1);
-        $times->attach($time2);
-        $times->attach($time3);
-        $times->attach($time4);
-
-        $event = new Event();
-        $event->setEventType('recurring');
-        $event->setDifferentTimes($times);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time3);
-        $expectedTimes->attach($time2);
-        $expectedTimes->attach($time1);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getSortedTimesForDate($event, $date)
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getSortedTimesForDateWithTimeExceptionsReturnsSortedTimes()
-    {
-        $date = new \DateTime('midnight');
-        $tomorrow = new \DateTime('midnight');
-        $tomorrow->modify('tomorrow');
-
-        $time1 = new Time();
-        $time1->setTimeBegin('23:56');
-        $time2 = new Time();
-        $time2->setTimeBegin('12:34');
-        $time3 = new Time();
-        $time3->setTimeBegin('00:34');
-        $time4 = new Time();
-        $time4->setTimeBegin('10:35');
-
-        $exception1 = new Exception();
-        $exception1->setExceptionType('Add');
-        $exception1->setExceptionDate($date);
-        $exception1->setExceptionTime($time1);
-        $exception2 = new Exception();
-        $exception2->setExceptionType('Time');
-        $exception2->setExceptionDate($date);
-        $exception2->setExceptionTime($time2);
-        $exception3 = new Exception();
-        $exception3->setExceptionType('Add');
-        $exception3->setExceptionDate($date);
-        $exception3->setExceptionTime($time3);
-        $exception4 = new Exception();
-        $exception4->setExceptionType('Add');
-        $exception4->setExceptionDate($tomorrow);
-        $exception4->setExceptionTime($time4);
-
-        $exceptions = new ObjectStorage();
-        $exceptions->attach($exception1);
-        $exceptions->attach($exception2);
-        $exceptions->attach($exception3);
-        $exceptions->attach($exception4);
-
-        $event = new Event();
-        $event->setExceptions($exceptions);
-
-        $expectedTimes = new \SplObjectStorage();
-        $expectedTimes->attach($time3);
-        $expectedTimes->attach($time2);
-        $expectedTimes->attach($time1);
-
-        self::assertEquals(
-            $expectedTimes,
-            $this->subject->getSortedTimesForDate($event, $date)
-        );
     }
 
     /**
